requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
aiohttp==2.3.2
async-timeout==2.0.0      # via aiohttp
certifi==2017.7.27.1
chardet==3.0.4            # via aiohttp
dawg-python==0.7.2        # via pymorphy2
docopt==0.6.2             # via pymorphy2
multidict==3.3.2          # via aiohttp, yarl
nltk==3.2.4
numpy==1.13.3             # via xgboost
pymorphy2-dicts==2.4.393442.3710985  # via pymorphy2
pymorphy2==0.8
pyyaml==3.12
schema==0.6.6
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0               # via nltk
wincertstore==0.2
xgboost==0.6
yarl==0.13.0              # via aiohttp
